feat(view): NavBarUserをRNに対応
